requirements.txt
opencv-python
mtcnn
keras-facenet
fastapi
uvicorn
scikit-learn
pillow
numpy
matplotlib
jupyter
